+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## v1.18.4
 
-[compare changes](https://github.com/Zotlo/zotlo-checkout/compare/v1.18.3...v1.18.4)
+[compare changes](https://github.com/Zotlo/zotlo-checkout/compare/v1.18.3...v1.18.5)
 
 ### 🩹 Fixes
 
-- **Z3-6542:** Capture SDK entry point errors in Sentry via Logger.wrap ([d20786a](https://github.com/Zotlo/zotlo-checkout/commit/d20786a))
+- **Z3-6542:** Capture SDK entry point errors in Sentry via Logger.wrap ([92073c7](https://github.com/Zotlo/zotlo-checkout/commit/92073c7))
+- **Z3-6542:** Handle Google Pay plain-object errors and guard partial config ([fb98297](https://github.com/Zotlo/zotlo-checkout/commit/fb98297))
 
 ### ❤️ Contributors